databento/03_schemas_and_symbology/requirements.txt
databento
python-dotenv
pandas
